R12: Social Listening - Friend activity, collaborative playlists, live sessions
- Add SocialMusicService with friend activity, follow system
- Add CollaborativePlaylistView for shared playlists with contributors
- Add LiveListeningSessionView with real-time reactions
- Add FriendActivityView for friend activity feed
- Fix pre-existing build errors in AIMusicService and ListeningInsightsView

+++ MuseMac/MuseMac/Views/PlaylistGeneratorView.swift
@@ -158,8 +158,8 @@ struct PlaylistGeneratorView: View {
     }
 
     private var estimatedDuration: Int {
-        let total = generatedPlaylist.reduce(0) { $0 + $1.duration }
-        return total / 60
+        let total = generatedPlaylist.reduce(0.0) { $0 + $1.duration }
+        return Int(total / 60)
     }
 
     // MARK: - Actions
